// web/src/app/types/types.ts
export interface Episode {
    _id: string;
    air_date: Date;
    duration: number;
    eid: number | string;
    full_titles: string[];
    title: string;
    anime_id: string;
    viewed: string[];
    view: boolean;
    iframe: string;
}

/** Track Timeline Interface
 * @property {String} id - Timeline id
 * @property {string} title - Timeline title
 * @property {Date} start - start date
 * @property {Date} end - End Date
 * @property {CSSProperties} style - HTML Style
 * @property {number} i - ?
 */
export interface TracTimeline{
    id: string;
    title: string;
    start: Date;
    end: Date;
    style: CSSProperties;
    i: number | string;
    element: IAnime
}

interface ISubTrack{
    id: string;
    title: string;
    elements: TracTimeline[]
}

/** ITrack Interface
 * @property {string} id - track id
 * @property {string} title - track title
 * @property {Array<TracTimeline>} elements - Subtracks
 * @property {boolean} isOpen - Open Track
 */
export interface ITrack{
    id: string;
    title: string;
    elements: TracTimeline[];
    tracks: ISubTrack[]
    isOpen: boolean;
}
// ...
export interface IAnime {
    _id: string;
    aid: number;
    bookmarks: string[];
    country?: string;
    episode: Episode[];
    episodes: number;
    title: string;
    full_title: string;
    image: string;
    title_en: string;
    title_jp?: string;
    quality: number;
    rating: number;
    release_date: Date;
    subscribe: boolean;
    subscribers?: string[];
    time?: number;
    url: string;
    description: string[];
    tags: string[]
}